% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raininfunction.R
\name{rainin}
\alias{rainin}
\title{This function allows you to take met station data and convert it from mm to in.}
\usage{
rainin(x)
}
\arguments{
\item{x}{a dataframe with one column of daily rainfall in units (mm)}
}
\value{
Rainfall data in units (in)
}
\description{
This function allows you to take met station data and convert it from mm to in.
}
